refactor: extract some fns into files
