+++ src/index.js
@@ -1,43 +1,11 @@
 import request from "./request/request.js";
-import { sorted, signed } from "./request/headers.js";
+import { sorted } from "./request/headers.js";
 import stringToSign from "./stringtosign.js";
 import signature from "./signature.js";
+import validate from "./validate.js";
+import authorization from "./authorization.js";
 
 const dateCleanRegex = /[:\-]|\.\d{3}/g;
-
-const requestRequired = [
-    "url",
-];
-
-const configRequired = [
-    "accessKeyId",
-    "region",
-    "secretAccessKey",
-    "service",
-];
-
-// Check for required params
-const validate = (source, config) => {
-    if(!source) {
-        throw new Error(`Missing request object`);
-    }
-
-    if(!config) {
-        throw new Error(`Missing config object`);
-    }
-
-    let missing = requestRequired.filter((field) => !source[field]);
-
-    if(missing.length) {
-        throw new Error(`Missing required request fields: ${missing.join(", ")}`);
-    }
-    
-    missing = configRequired.filter((field) => !config[field]);
-
-    if(missing.length) {
-        throw new Error(`Missing required config fields: ${missing.join(", ")}`);
-    }
-};
 
 const parseDate = ({ headers }) => {
     const datetime = "X-Amz-Date" in headers ?
@@ -50,16 +18,6 @@ const parseDate = ({ headers }) => {
         short : datetime.split("T")[0],
         long  : datetime,
     };
-};
-
-const authorization = (req, sig) => {
-    const { algorithm, accessKeyId, date, region, service, sortedHeaders } = req;
-
-    return [
-        `${algorithm} Credential=${accessKeyId}/${date.short}/${region}/${service}/aws4_request`,
-        `SignedHeaders=${signed(sortedHeaders)}`,
-        `Signature=${sig}`,
-    ].join(", ");
 };
 
 export default (source, config) => {
@@ -83,7 +41,7 @@ export default (source, config) => {
             sortedHeaders : sorted(source),
         }
     );
-    
+
     const canonical = request(details);
     const sts = stringToSign(details, canonical);
     const sig = signature(details, sts);
